Flyttet funksjonen kortNedBeskrivelse til en egen fil. Laget tester til funksjonen.
